+++ domain/src/main/java/com/keyme/domain/usecase/GetMySolvedScoreUseCase.kt
@@ -1,12 +1,17 @@
 package com.keyme.domain.usecase
 
 import com.keyme.domain.entity.apiResult
-import com.keyme.domain.entity.response.SolvedScore
+import com.keyme.domain.entity.response.QuestionSolvedScore
 import com.keyme.domain.repository.QuestionRepository
 import javax.inject.Inject
 
 class GetMySolvedScoreUseCase @Inject constructor(
     private val questionRepository: QuestionRepository,
 ) {
-    operator fun invoke() = SolvedScore("", 3)
+    suspend operator fun invoke(questionId: String) =
+        apiResult {
+            questionRepository.getSolvedScore(
+            questionId = questionId,
+            ownerId = "1")
+        }
 }
